Added status class
